+++ src/pack/make_pack_descriptor.hpp
@@ -13,104 +13,76 @@
 #ifndef PACK_MAKE_PACK_DESCRIPTOR_HPP_
 #define PACK_MAKE_PACK_DESCRIPTOR_HPP_
 
-#include <algorithm>
-#include <functional>
-#include <limits>
-#include <map>
 #include <memory>
-#include <regex>
 #include <set>
 #include <string>
 #include <tuple>
-#include <type_traits>
+#include <unordered_map>
 #include <utility>
 #include <vector>
 
-#include "interface/mesh_data.hpp"
-#include "interface/meshblock_data.hpp"
 #include "interface/metadata.hpp"
 #include "interface/state_descriptor.hpp"
-#include "interface/variable.hpp"
 #include "mesh/mesh.hpp"
+#include "pack/pack_descriptor.hpp"
 #include "pack/sparse_pack.hpp"
 #include "utils/type_list.hpp"
 
 namespace parthenon {
+namespace impl {
+PackDescriptor MakePackDescriptorBase(StateDescriptor *psd,
+                                      const std::vector<std::string> &vars,
+                                      const std::vector<bool> &use_regex,
+                                      const std::vector<MetadataFlag> &flags,
+                                      const std::set<PDOpt> &options);
+PackDescriptor MakePackDescriptorBase(StateDescriptor *psd,
+                                      const std::vector<Uid_t> &var_ids,
+                                      const std::vector<MetadataFlag> &flags,
+                                      const std::set<PDOpt> &options);
+template <class MT>
+StateDescriptor *GetStateDescriptor(MT *pmd);
+} // namespace impl
 
 inline auto MakeDefaultPackDescriptor() { return typename SparsePack<>::Descriptor(); }
 
-inline auto MakePackDescriptor(StateDescriptor *psd, const std::vector<std::string> &vars,
+template <class MT>
+inline auto MakePackDescriptor(MT *pmd, const std::vector<std::string> &vars,
                                const std::vector<bool> &use_regex,
                                const std::vector<MetadataFlag> &flags = {},
                                const std::set<PDOpt> &options = {}) {
-  PARTHENON_REQUIRE(vars.size() == use_regex.size(),
-                    "Vargroup names and use_regex need to be the same size.");
-  auto selector = [&](int vidx, const VarID &id, const Metadata &md) {
-    if (flags.size() > 0) {
-      for (const auto &flag : flags) {
-        if (!md.IsSet(flag)) return false;
-      }
-    }
-
-    if (use_regex[vidx]) {
-      if (std::regex_match(std::string(id.label()), std::regex(vars[vidx]))) return true;
-    } else {
-      if (vars[vidx] == id.label()) return true;
-      if (vars[vidx] == id.base_name && id.sparse_id != InvalidSparseID) return true;
-    }
-    return false;
-  };
-
-  impl::PackDescriptor base_desc(psd, vars, selector, options);
-  return typename SparsePack<>::Descriptor(base_desc);
+  return typename SparsePack<>::Descriptor(impl::MakePackDescriptorBase(
+      impl::GetStateDescriptor<MT>(pmd), vars, use_regex, flags, options));
 }
 
-template <class... Ts>
-inline auto MakePackDescriptor(StateDescriptor *psd,
+template <class MT>
+inline auto MakePackDescriptor(MT *pmd, const std::vector<std::string> &vars,
                                const std::vector<MetadataFlag> &flags = {},
                                const std::set<PDOpt> &options = {}) {
-  static_assert(sizeof...(Ts) > 0, "Must have at least one variable type for type pack");
-
-  std::vector<std::string> vars{Ts::name()...};
-  std::vector<bool> use_regex{Ts::regex()...};
-
-  return typename SparsePack<Ts...>::Descriptor(static_cast<impl::PackDescriptor>(
-      MakePackDescriptor(psd, vars, use_regex, flags, options)));
-}
-
-inline auto MakePackDescriptor(StateDescriptor *psd, const std::vector<std::string> &vars,
-                               const std::vector<MetadataFlag> &flags = {},
-                               const std::set<PDOpt> &options = {}) {
-  return MakePackDescriptor(psd, vars, std::vector<bool>(vars.size(), false), flags,
+  return MakePackDescriptor(pmd, vars, std::vector<bool>(vars.size(), false), flags,
                             options);
 }
 
-template <class... Ts>
-inline auto MakePackDescriptor(MeshBlockData<Real> *pmbd,
-                               const std::vector<MetadataFlag> &flags = {},
+template <class... Ts, class MT>
+inline auto MakePackDescriptor(MT *pmd, const std::vector<MetadataFlag> &flags = {},
                                const std::set<PDOpt> &options = {}) {
-  return MakePackDescriptor<Ts...>(
-      pmbd->GetBlockPointer()->pmy_mesh->resolved_packages.get(), flags, options);
+  const std::vector<std::string> vars{Ts::name()...};
+  const std::vector<bool> use_regex{Ts::regex()...};
+  return typename SparsePack<Ts...>::Descriptor(static_cast<impl::PackDescriptor>(
+      MakePackDescriptor(pmd, vars, use_regex, flags, options)));
 }
 
-template <class... Ts>
-inline auto MakePackDescriptor(MeshData<Real> *pmd,
+template <class... Ts, class MT>
+inline auto MakePackDescriptor(SparsePack<Ts...> pack, MT *pmd,
                                const std::vector<MetadataFlag> &flags = {},
                                const std::set<PDOpt> &options = {}) {
-  return MakePackDescriptor<Ts...>(pmd->GetMeshPointer()->resolved_packages.get(), flags,
-                                   options);
+  return parthenon::MakePackDescriptor<Ts...>(pmd, flags, options);
 }
 
-template <class... Ts>
-inline auto MakePackDescriptor(SparsePack<Ts...> pack, StateDescriptor *psd,
-                               const std::vector<MetadataFlag> &flags = {},
-                               const std::set<PDOpt> &options = {}) {
-  return parthenon::MakePackDescriptor<Ts...>(psd, flags, options);
-}
-
-inline auto MakePackDescriptor(
-    StateDescriptor *psd, const std::vector<std::pair<std::string, bool>> &var_regexes,
-    const std::vector<MetadataFlag> &flags = {}, const std::set<PDOpt> &options = {}) {
+template <class MT>
+inline auto
+MakePackDescriptor(MT *psd, const std::vector<std::pair<std::string, bool>> &var_regexes,
+                   const std::vector<MetadataFlag> &flags = {},
+                   const std::set<PDOpt> &options = {}) {
   std::vector<std::string> vars;
   std::vector<bool> use_regex;
   for (const auto &[v, r] : var_regexes) {
@@ -120,21 +92,12 @@ inline auto MakePackDescriptor(
   return MakePackDescriptor(psd, vars, use_regex, flags, options);
 }
 
-inline auto MakePackDescriptor(StateDescriptor *psd, const std::vector<Uid_t> &var_ids,
+template <class MT>
+inline auto MakePackDescriptor(MT *pmd, const std::vector<Uid_t> &var_ids,
                                const std::vector<MetadataFlag> &flags = {},
                                const std::set<PDOpt> &options = {}) {
-  auto selector = [&](int vidx, const VarID &id, const Metadata &md) {
-    if (flags.size() > 0) {
-      for (const auto &flag : flags) {
-        if (!md.IsSet(flag)) return false;
-      }
-    }
-    if (Variable<Real>::GetUniqueID(id.label()) == var_ids[vidx]) return true;
-    return false;
-  };
-
-  impl::PackDescriptor base_desc(psd, var_ids, selector, options);
-  return typename SparsePack<>::Descriptor(base_desc);
+  return typename SparsePack<>::Descriptor(impl::MakePackDescriptorBase(
+      impl::GetStateDescriptor<MT>(pmd), var_ids, flags, options));
 }
 
 template <template <class...> class TL, class... Types, class... Args>
@@ -146,6 +109,49 @@ template <class TL, class... Args>
 inline auto MakePackDescriptorFromTypeList(Args &&...args) {
   return MakePackDescriptorFromTypeList(TL(), std::forward<Args>(args)...);
 }
+
+struct PackDescriptorCacheBase {
+  virtual ~PackDescriptorCacheBase() = default;
+};
+
+template <class... Ts>
+class PackDescCache : public PackDescriptorCacheBase {
+ public:
+  using key_t = std::tuple<Ts...>;
+  std::unordered_map<key_t, impl::PackDescriptor> map;
+
+  static std::optional<impl::PackDescriptor>
+  CheckForKey(StateDescriptor *pdesc, const std::string &cache_label, const Ts &...args) {
+    if (pdesc) {
+      if (!pdesc->pack_desc_cache_map.count(cache_label)) {
+        // Create a cache for PackDescriptors created with this particular selector
+        auto pcache = std::make_shared<PackDescCache>();
+        pdesc->pack_desc_cache_map.emplace(
+            cache_label, std::dynamic_pointer_cast<PackDescriptorCacheBase>(pcache));
+      } else {
+        // Check if a PackDescriptor already exists in the cache
+        auto pcache = std::dynamic_pointer_cast<PackDescCache>(
+            pdesc->pack_desc_cache_map[cache_label]);
+        auto key = std::make_tuple(args...);
+        if (pcache->map.count(key))
+          return std::optional<impl::PackDescriptor>{pcache->map[key]};
+      }
+    }
+    return std::nullopt;
+  }
+
+  static void CachePackDescriptor(StateDescriptor *pdesc, const std::string &cache_label,
+                                  const impl::PackDescriptor &pd, const Ts &...args) {
+    if (pdesc) {
+      // Store the newly created PackDescriptor in the cache
+      auto pcache = std::dynamic_pointer_cast<PackDescCache>(
+          pdesc->pack_desc_cache_map[cache_label]);
+      auto key = std::make_tuple(args...);
+      pcache->map.emplace(key, pd);
+    }
+  }
+};
+
 } // namespace parthenon
 
 #endif // PACK_MAKE_PACK_DESCRIPTOR_HPP_